# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Defines Canadian Insurance-Mortgage POC REST services"
  version: "1.0.0"
  title: "Canadian Insurance-Mortgage"
  contact:
    email: "kumari.shweta@hcl.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "@@host-details@@"
basePath: "/api"
...
schemes:
- "http"
paths:
  /authenticate:
    options:
      tags:
      - "Authentication"
      summary: "Authenticate user into the system"
      description: "This service is used to Authenticate user"
      operationId: "optionsMethodAuth"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "bad request"
        404:
          description: "not found"
      x-swagger-router-controller: "userController" 
    post:
      tags:
      - "Authentication"
      summary: "Authenticate user into the system"
      description: "This service is used to Authenticate user"
      operationId: "authenticateUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "user object that needs for authenticate"
        required: true
        schema:
          $ref: "#/definitions/user"
      responses:
        405:
          description: "Invalid input"
      x-swagger-router-controller: "userController"  
  /validateToken:
    options:
      tags:
      - "Authentication"
      summary: "Get success for options"
      description: "This service used to send 200 for option method call"
      operationId: "optionsMethodValidateToken"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "bad request"
        404:
          description: "not found"
      x-swagger-router-controller: "userController"
    post:
      tags:
      - "Authentication"
      summary: "Get the token validated"
      description: "This service is used to validate teh token"
      operationId: "validateToken"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Token to be added"
        required: true
        schema:
          $ref: "#/definitions/validateTokenForPost"
      responses:
        200:
          description: "Token is valid"
        405:
          description: "Invalid Token supplied"
      x-swagger-router-controller: "userController" 
  /findByPersona:
    options:
      tags:
      - "Authentication"
      summary: "Get success for options"
      description: "This service used to send 200 for option method call"
      operationId: "optionsMethodFindPersona"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "bad request"
        404:
          description: "not found"
      x-swagger-router-controller: "userController" 
    get:
      tags:
      - "Authentication"
      summary: "Get user information by persona"
      description: "This service is used to get user information by persona"
      operationId: "findByPersona"
      security:
        - Bearer: []
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: false
        type: "string"
      - name: "persona"
        in: "query"
        description: "PersonaDetails "
        required: true
        type: "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/getuserobject"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Audit not found"
      x-swagger-router-controller: "userController" 
  /findByUserId:
    options:
      tags:
      - "Authentication"
      summary: "Get success for options"
      description: "This service used to send 200 for option method call"
      operationId: "optionsMethodFindByUserId"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
        400:
          description: "bad request"
        404:
          description: "not found"
      x-swagger-router-controller: "userController"
    get:
      tags:
      - "Authentication"
      summary: "Get user information by userId"
      description: "This service is used to get user information by userId"
      operationId: "findByUserId"
      security:
        - Bearer: []
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: false
        type: "string"
      - name: "userId"
        in: "query"
        description: "userId "
        required: true
        type: "string"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/getuserobject"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Audit not found"
      x-swagger-router-controller: "userController"  
...
securityDefinitions:
  dlpstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:ipnotice: "write IP notice"
      read:ipnotice: "read IP notice"
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"  
definitions:
...
  validateTokenForPost:
    type: "object"
    properties:
      token:
        type: "string"
    xml:
      name: "Block" 
...
  getuserobject:
    type: "object"
    properties:
      statusCode:
        type: "string"
      result:
        type: "array"
        items:
          $ref: "#/definitions/user"
    example:
      statusCode: 200
      result: [{
       userId: "userId",
       userName: "userName",
       createdDate: "createdDate",
       createdBy: "createdBy",
       persona: "persona"
      }]
    xml:
      name: "getuserobject" 
  user:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      userId:
        type: "string"
      password:
        type: "string" 
      subscriptionKey:
        type: "string"
    example:
       userId: "userId"
       password: "password"
       subscriptionKey: "subscriptionKey"
    xml:
      name: "user"     